customize the layout macros
the original macros have every possible switch position for the board,
but there are a handful of switch positions that I'm not using. It's not
actually possible to use them all at once, since some of them would
cause keys to collide. I removed the uneeded positions to fit my layout.